requirements.txt
wheel==0.37.1
nltk==3.4.5
pymongo==3.12.0
gensim==4.1.2
flask==2.0.3
scikit-learn==1.0.2
PyYaml==6.0
Flask-Cors==3.0.10